pyproject.toml
[project]
name = "prisminspire"
version = "0.1.0"
description = ""
authors = [
    {name = "Yajnesh S Kottary",email = "yajnesh.k@pacewisdom.com"}
]
requires-python = ">=3.14,<4.0"
dependencies = [
    "alembic (>=1.18.1,<2.0.0)",
    "async-openai (>=0.0.52,<0.0.53)",
    "fastapi (>=0.128.0,<0.129.0)",
    "fastapi-utils (>=0.8.0,<0.9.0)",
    "async-timeout (>=5.0.1,<6.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "boto3 (>=1.42.34,<2.0.0)",
    "chardet (>=5.2.0,<6.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "email-validator (>=2.3.0,<3.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "httptools (>=0.7.1,<0.8.0)",
    "langchain (>=1.2.7,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
    "langchain-milvus (>=0.3.3,<0.4.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "mangum (>=0.20.0,<0.21.0)",
    "markdown2 (>=2.5.4,<3.0.0)",
    "mcp (>=1.26.0,<2.0.0)",
    "milvus-lite (>=2.5.1,<3.0.0)",
    "msoffcrypto-tool (>=6.0.0,<7.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pypdf (>=6.6.0,<7.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-toon (>=0.1.3,<0.2.0)",
    "reportlab (>=4.4.9,<5.0.0)",
    "ujson (>=5.11.0,<6.0.0)",
    "uvloop (>=0.22.1,<0.23.0)",
    "watchfiles (>=1.1.1,<2.0.0)",
    "google-genai (>=1.60.0,<2.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "pytz (>=2025.2,<2026.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pipdeptree (>=2.30.0,<3.0.0)"
]